requirements.txt
fastapi
openai
python-dotenv
pymongo
argon2-cffi
PyJWT
uvicorn